Refactor color scale out of map legend
