+++ client/src/components/stats/AllMountains.tsx
@@ -1,4 +1,5 @@
 import { useQuery } from '@apollo/react-hooks';
+import { faMapMarkedAlt } from '@fortawesome/free-solid-svg-icons';
 import { GetString } from 'fluent-react/compat';
 import gql from 'graphql-tag';
 import noop from 'lodash/noop';
@@ -8,6 +9,8 @@ import {
   AppLocalizationAndBundleContext,
 } from '../../contextProviders/getFluentLocalizationContext';
 import {
+  BasicIconInText,
+  DetailBoxTitle,
   PlaceholderText,
   SelectBox,
 } from '../../styling/styleUtils';
@@ -16,21 +19,12 @@ import {
   PeakListVariants,
   User,
 } from '../../types/graphQLTypes';
-import {
-  failIfValidOrNonExhaustive,
-} from '../../Utils';
 import getCompletionDates, {VariableDate} from '../peakLists/detail/getCompletionDates';
 import MountainTable from '../peakLists/detail/MountainTable';
+import { getColorScale } from '../peakLists/detail/PeakListDetail';
 import LoadingSpinner from '../sharedComponents/LoadingSpinner';
 import Map, {MapContainer} from '../sharedComponents/map';
-import {
-  fiveColorScale,
-  fiveSymbolScale,
-  thirteenColorScale,
-  thirteenSymbolScale,
-  twoColorScale,
-  twoSymbolScale,
-} from '../sharedComponents/map/colorScaleColors';
+import MountainColorScale from '../sharedComponents/map/MountainColorScale';
 import Header from './Header';
 
 const SelectBoxContiner = styled.div`
@@ -132,41 +126,9 @@ const AllMountains = (props: Props) => {
       );
     } else {
 
-      let colorScaleTitle: string | undefined;
-      let colorScaleColors: string[];
-      let colorScaleSymbols: string[];
-      let colorScaleLabels: string[];
-      if (type === PeakListVariants.standard || type === PeakListVariants.winter) {
-        colorScaleTitle = undefined;
-        colorScaleColors = twoColorScale;
-        colorScaleSymbols = twoSymbolScale;
-        colorScaleLabels = [
-          getFluentString('global-text-value-not-done'),
-          getFluentString('global-text-value-done'),
-        ];
-      } else if (type === PeakListVariants.fourSeason) {
-        colorScaleTitle = getFluentString('map-number-of-seasons');
-        colorScaleColors = fiveColorScale;
-        colorScaleSymbols = fiveSymbolScale;
-        colorScaleLabels = [
-          getFluentString('map-no-seasons'),
-          getFluentString('map-all-seasons'),
-        ];
-      } else if (type === PeakListVariants.grid) {
-        colorScaleTitle = getFluentString('map-number-of-months');
-        colorScaleColors = thirteenColorScale;
-        colorScaleSymbols = thirteenSymbolScale;
-        colorScaleLabels = [
-          getFluentString('map-no-months'),
-          getFluentString('map-all-months'),
-        ];
-      } else {
-        colorScaleTitle = undefined;
-        colorScaleColors = [];
-        colorScaleSymbols = [];
-        colorScaleLabels = [];
-        failIfValidOrNonExhaustive(type, 'Invalid peak list type ' + type);
-      }
+      const {
+        colorScaleTitle, colorScaleColors, colorScaleSymbols, colorScaleLabels,
+      } = getColorScale(type, getFluentString);
 
       const userMountains = mountains ? mountains : [];
 
@@ -242,16 +204,25 @@ const AllMountains = (props: Props) => {
           />
           {toggleType}
           <MapContainer>
+
+            <DetailBoxTitle>
+              <BasicIconInText icon={faMapMarkedAlt} />
+              {getFluentString('map-generic-title')}
+            </DetailBoxTitle>
+            <MountainColorScale
+              colorScaleColors={colorScaleColors}
+              colorScaleSymbols={colorScaleSymbols}
+              colorScaleLabels={colorScaleLabels}
+              colorScaleTitle={colorScaleTitle}
+            />
             <Map
               peakListId={null}
               mountainId={'all'}
               coordinates={allMountainsWithDates}
               highlighted={undefined}
               userId={userId}
-              colorScaleTitle={colorScaleTitle}
               colorScaleColors={colorScaleColors}
               colorScaleSymbols={colorScaleSymbols}
-              colorScaleLabels={colorScaleLabels}
               completedAscents={userMountains}
               key={'stats-all-mountains-in-progress-and-complete-key'}
             />
